pyproject.toml
[project]
name = "mcp-database-query"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.13.0.2",
    "mysql-connector-python>=9.5.0",
    "psycopg2-binary>=2.9.11",
    "pymongo>=4.15.3",
    "python-dotenv>=1.2.1",
]
